Add forgotten change in the integration tests.
git-svn-id: https://trac.musala.com/ARDC/repo@10022 34708651-a5da-c34d-b20c-10ddab38317d

+++ test/com/musala/atmosphere/client/device/WaitForExistsTest.java
@@ -21,7 +21,7 @@ import com.musala.atmosphere.commons.cs.clientbuilder.DeviceParameters;
 
 public class WaitForExistsTest extends BaseIntegrationTest {
 
-    private static final Long ELEMENT_WAIT_TIMEOUT = 10000l;
+    private static final Integer ELEMENT_WAIT_TIMEOUT = 10000;
 
     private static final String GESTURE_ACTIVITY_BUTTON_DESCRIPTOR = "GestureActivityButton";
